Relocate views.sql to packages/civitai-db-schema/prisma/programmability/ (monorepo path read by prisma-prepare-programmability.mjs) and drop ImageStat/ImageRank_Live, which referenced the ImageMetric table removed in #2881 and would fail db:program's single-transaction apply. Resolve update_model_rank.sql deletion against main's relocated copy.

+++ src/utils/metadata/index.ts
@@ -28,7 +28,7 @@ export async function ExifParser(file: File | string) {
 
   if (exif.UserComment) {
     // @ts-ignore - this is a hack to not have to rework our downstream code
-    exif.userComment = Int32Array.from(exif.UserComment);
+    exif.userComment = Uint8Array.from(exif.UserComment);
   }
 
   // Clone exif before each canParse to prevent cross-parser mutation
